[Feat] 기능 추가
무기 구매 기능 추가

+++ Assets/Scripts/Player/Player.cs
@@ -30,7 +30,7 @@ public class Player : MonoBehaviour
     public TextMeshProUGUI goldAmountText;
     public TextMeshProUGUI goldCostText;
 
-    public int curgold;
+    public float curgold;
 
     public void Start()
     {
